404-world-not-found/src/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-PyMongo==3.0.1
Flask-WTF==1.2.2
google-auth==2.38.0
google-genai==1.10.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mongoengine==0.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.0
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.2.1
